requirements.txt
flask
tensorflow
Pillow
numpy
matplotlib
scikit-learn